Create features module
